Refactored namespace structure
