+++ ArchitectureParserTest/ConnectionTest.cs
@@ -2,7 +2,7 @@
 
 using ArchitectureParser.Architecture.Components;
 using ArchitectureParser.Architecture.Factories;
-using ArchitectureParser.Architecture.NullObjects;
+using ArchitectureParser.Architecture.Connections;
 
 namespace ArchitectureParserTest
 {
